chore(lint): hold the vendored squashfs and erofs forks to repo lint rules
Drops the blanket path exclusions for internal/squashfs and internal/erofs
and fixes what they were hiding (~170 findings): Go-style names for the
superblock flag and stat constants and ID initialisms, receiver-naming and
import-shadowing cleanups (via gopls rename), slices.SortFunc migrations,
explicit defaults on on-disk type switches, comment/increment style, a
shared writeMetaBlocks helper replacing duplicated metadata-block emission,
always-nil error returns removed from flushBlock and writeDirEntries, and
flattened match-search and trailing-data paths in the erofs writer.

What remains excluded is narrow and documented in .golangci.yml following
the existing per-path rule pattern: complexity metrics (maintidx, nestif)
on the forks' upstream-inherited parser and tree-copy functions, and
gochecknoglobals on the codec registry and xattr lookup tables.

Since the touched code emits on-disk formats, both squashfs and erofs
images were re-validated after the refactor with fsck, kernel mounts, and
recursive content diffs.

+++ internal/squashfs/hardlink_unix.go
@@ -13,5 +13,5 @@ func getDevIno(sys any) (devIno, bool) {
 	if !ok {
 		return devIno{}, false
 	}
-	return devIno{dev: uint64(st.Dev), ino: uint64(st.Ino)}, true
+	return devIno{dev: uint64(st.Dev), ino: st.Ino}, true
 }
